code/optional_requirements.txt
scikit_learn>=1.3.0 #partitioning
cwb-ccc>=0.12.4 #CWB
openai>=1.40.0 #commercial
pydantic>=1.8.2 #commercial
requests>=2.20.0 #web
PyICU>=2.11 #ICU #sorting
spacy>=3.8.4 #annotation
sentence_transformers>=2.1.0 #annotation
ipywidgets>=7.6.5 #notebooks
wordfreq>=2.8.0 #frequencies
